+++ master_src/least_difference.go
@@ -10,7 +10,7 @@ type LeastDifference struct {
 }
 
 // TODO - need locks here?
-func (l *LeastDifference) assignTask(load int) (*Slave, error) {
+func (l *LeastDifference) assignTask(load uint64) (*Slave, error) {
 	// TODO - need to lock Slavepool!!!
 	l.slavePool.mtx.RLock()
 	defer l.slavePool.mtx.RUnlock()
